requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-etc==1.2.0
django-extensions==3.0.9
django-hitcount==1.3.3
django-js-asset==1.2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Faker==6.4.1
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
